Updated for style and out of source builds

+++ build.sh
@@ -2,7 +2,7 @@
 . /etc/profile.d/modules.sh
 SOURCE_FILE=${NAME}-${VERSION}.tar.gz
 
-module load ci
+module add ci
 module add gmp
 module add mpfr
 module add ncurses
@@ -26,16 +26,27 @@ mkdir -p ${SOFT_DIR}
 
 #  Download the source file
 
-if [[ ! -s ${SRC_DIR}/${SOURCE_FILE} ]] ; then
+if [ ! -e ${SRC_DIR}/${SOURCE_FILE}.lock ] && [ ! -s ${SRC_DIR}/${SOURCE_FILE} ] ; then
+  touch  ${SRC_DIR}/${SOURCE_FILE}.lock
   echo "seems like this is the first build - let's get the source"
   mkdir -p ${SRC_DIR}
   wget http://mirror.ufs.ac.za/gnu/gnu/${NAME}/${SOURCE_FILE} -O ${SRC_DIR}/${SOURCE_FILE}
-  tar xvf ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE}
-else
+  echo "releasing lock"
+  rm -v ${SRC_DIR}/${SOURCE_FILE}.lock
+  elif [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; then
+  # Someone else has the file, wait till it's released
+  while [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; do
+    echo " There seems to be a download currently under way, will check again in 5 sec"
+    sleep 5
+  done
+  else
   echo "continuing from previous builds, using source at " ${SRC_DIR}/${SOURCE_FILE}
-  tar xvzf ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE} --skip-old-files
 fi
-cd ${WORKSPACE}/${NAME}-${VERSION}
+mkdir -p ${WORKSPACE}/${NAME}-${VERSION}/build-${BUILD_NUMBER}
+cd ${WORKSPACE}/${NAME}-${VERSION}/build-${BUILD_NUMBER}
 # We could probably loop through the dependencies here
-./configure --prefix ${SOFT_DIR} --with-mpfr=${MPFR_DIR}  --with-gmp=${GMP_DIR}
-make -j 8
+../configure \
+--prefix=${SOFT_DIR} \
+--with-mpfr=${MPFR_DIR} \
+--with-gmp=${GMP_DIR}
+make -j 2
